inputhook: make stdin_ready() function reusable
Move it from lib.inputhookwx to lib.inputhook, as the former
module can only be loaded if wx itself is available.
